+++ Sources/App/Services/Accounts/SessionService.swift
@@ -12,7 +12,9 @@ struct SessionService {
 
     /// Fetches sessions belonging to an account.
     func fetchSessions() async throws -> [Session] {
-        let account = try request.authService.getUser().account
+        guard let account = try request.authService.getUser().account else {
+            throw Abort(.internalServerError, reason: "Could not process your account. Try again in a little bit.")
+        }
         return try await account.$sessions.get(on: request.db)
     }
 
@@ -65,7 +67,9 @@ struct SessionService {
 
     /// Deletes a given session via its ID.
     func deleteSession(_ id: UUID) async throws {
-        let account = try request.authService.getUser().account
+        guard let account = try request.authService.getUser().account else {
+            throw Abort(.internalServerError, reason: "Could not process your account. Try again in a little bit.")
+        }
         return try await account.$sessions.query(on: request.db)
             .filter(\.$id == id)
             .delete()
